Create Dell-Command-Update.py
I have created a script to automate downloading "Dell Command Update". However, after successfully downloading it. I am running into the issue of when clicking on the file. It states "the file or directory is corrupted and unreadable". If anyone knows how to solve this problem, it would be greatly appreciated.